requirements.txt

fastapi>=0.122.0
httpx>=0.28.1
pydantic>=2.12.5
python-dotenv>=1.2.1
pyyaml>=6.0.3
uvicorn>=0.38.0
fastapi
httpx
pydantic
python-dotenv
pyyaml
uvicorn
python-multipart
aiofiles
asyncpg
bcrypt
passlib
python-jose
python-multipart
redis
sqlalchemy
